+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KF9VM0EXCWENTFHPAC5BNAGJ"
+        "01KFAQP8AWWF83VG4RGG7FP4PZ"
     )
     r"""The unique identifier of the data source"""