+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KF9VM0EWR6DAG9T14ZNX8G8D"
+        "01KFAQP8AVFENG5ZBXXK4PFBRG"
     )
     r"""The unique identifier of the data source"""