+++ packages/orq-rc/src/orq_ai_sdk/models/updateidentityop.py
@@ -194,7 +194,7 @@ class UpdateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-19T00:48:48.487Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-19T08:59:21.884Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")